+++ api-module/src/main/java/org/devridge/api/domain/community/dto/response/UserInformation.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @Getter
 @Builder
 @AllArgsConstructor
-public class MemberInfoResponse {
+public class UserInformation {
 
     @JsonProperty("id")
     private Long memberId;
