+++ api-module/src/main/java/org/devridge/api/domain/note/dto/response/ReceivedParticipationNoteListResponse.java
@@ -3,13 +3,13 @@ package org.devridge.api.domain.note.dto.response;
 import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Getter;
-import org.devridge.api.domain.community.dto.response.MemberInfoResponse;
+import org.devridge.api.common.dto.UserInformation;
 
 @Getter
 @Builder
 public class ReceivedParticipationNoteListResponse {
 
-    private MemberInfoResponse sendMember;
+    private UserInformation sendMember;
     private LocalDateTime receivedTime;
     private Boolean isApproved;
 }
