+++ api-module/src/main/java/org/devridge/api/domain/note/dto/response/SentParticipationNoteListResponse.java
@@ -3,13 +3,13 @@ package org.devridge.api.domain.note.dto.response;
 import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Getter;
-import org.devridge.api.domain.community.dto.response.MemberInfoResponse;
+import org.devridge.api.common.dto.UserInformation;
 
 @Getter
 @Builder
 public class SentParticipationNoteListResponse {
 
-    private MemberInfoResponse receiveMember;
+    private UserInformation receiveMember;
     private LocalDateTime sentTime;
     private Boolean isApproved;
 }
