+++ api-module/src/main/java/org/devridge/api/domain/note/entity/ProjectParticipationNote.java
@@ -11,9 +11,9 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.devridge.api.common.entity.BaseEntity;
 import org.devridge.api.domain.community.entity.Project;
 import org.devridge.api.domain.member.entity.Member;
-import org.devridge.common.entity.BaseEntity;
 import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.DynamicInsert;
