+++ api-module/src/main/java/org/devridge/api/domain/note/exception/NoteForbiddenException.java
@@ -1,6 +1,7 @@
 package org.devridge.api.domain.note.exception;
 
-import org.devridge.api.exception.common.BaseException;
+
+import org.devridge.api.common.exception.common.BaseException;
 
 public class NoteForbiddenException extends BaseException {
