+++ api-module/src/main/java/org/devridge/api/domain/note/exception/ParticipationNoteForbiddenException.java
@@ -1,6 +1,6 @@
 package org.devridge.api.domain.note.exception;
 
-import org.devridge.api.exception.common.BaseException;
+import org.devridge.api.common.exception.common.BaseException;
 
 public class ParticipationNoteForbiddenException extends BaseException {
